build(all): use es2019 target to support on Node12 (#925)
Switching to target 2019 to support Node12

Previously, we have been building our packages on target es2020. When customers consume our modules, it will still have ?.. If customers use NodeJS12 (which is es2019), it will throw an error SyntaxError: Unexpected token '.'. The . is the second character of ?. syntax.

To fix this, we change all of the target to 2019. However, this cause a lot of breaks in our current code.

1. Branch coverage will drop as ?. is broken into a conditional statement. We need to add new test cases or tell TS that the value does exist by adding !. This is really case by case as the value cannot really be undefined in some case.
2. NodeJS12 doesn't have crypto.randomUUID that we use in e2e tests. So we have to add uuid as a devDependency

This commit also adds the additional checks on NodeJS12 and NodeJS14. (We tested only on Node16). The failing one is Node12, and it fails on examples/sam package. This package uses the npm published version of our library. And it really should fail as the currently published library is target 2020. Once we publish a new version, this should be fixed automatically.